Update dnq project.py

+++ dnq project.py
@@ -4,35 +4,42 @@ print(20*'=')
 print(' DIVIDE AND CONQUER')
 print(20*'=')
 
-x = int(input('maximum = '))
-num_queue = queue.Queue(maxsize = x)
-num_arr = []
+max_number = int(input('How many numbers do we have = '))
 
-for a in range(0,x):
-    num_arr.append(int(input('Input numbers = ')))
+if(max_number < 5):
+    print('Input more numbers!')
+    print('This program works for at least 5 numbers.')
 
-print('')
-print(35*'-')
-print(' ARRAY [BEFORE DIVIDE AND CONQUER]')
-print(35*'-')
-print(num_arr)
-
-num_queue.put(num_arr)
-arr_mid = []
-
-while(num_queue.qsize() > 0):
-    arr_check = num_queue.get() #get from queue and delete from it
-    if len(arr_check) > 0:
-        mid_idx = len(arr_check)//2 #// floor division
-        mid = arr_check[mid_idx]
-        arr_left = arr_check[0:mid_idx]
-        arr_right = arr_check[mid_idx+1:]
-        arr_mid.append(mid)
-        num_queue.put(arr_left)
-        num_queue.put(arr_right)
+else:
+    num_queue = queue.Queue(maxsize = max_number)
+    num_arr = []
     
-print('')
-print(34*'-')
-print(' ARRAY [AFTER DIVIDE AND CONQUER]')
-print(34*'-')  
-print(arr_mid)
+    for n in range(0,max_number):
+        num_input = int(input('Input numbers [' + str(n+1) + '] = '))
+        num_arr.append(num_input)
+    
+    print('')
+    print(35*'-')
+    print(' ARRAY [BEFORE DIVIDE AND CONQUER]')
+    print(35*'-')
+    print(num_arr) 
+    
+    num_queue.put(num_arr)
+    arr_mid = []
+    
+    while(num_queue.qsize() > 0):
+        arr_check = num_queue.get() #get from queue and delete from it
+        if len(arr_check) > 0:
+            mid_idx = len(arr_check)//2 #// floor division
+            mid = arr_check[mid_idx]
+            arr_left = arr_check[0:mid_idx]
+            arr_right = arr_check[mid_idx+1:]
+            arr_mid.append(mid)
+            num_queue.put(arr_left)
+            num_queue.put(arr_right)
+        
+    print('')
+    print(34*'-')
+    print(' ARRAY [AFTER DIVIDE AND CONQUER]')
+    print(34*'-')  
+    print(arr_mid)
